No more jQuery, added some helper functions for DOM visibility
The library is now completely independent of jQuery. The interface is also independent, but of course Bootstrap still needs jQuery. Maybe I should get rid of Bootstrap and jQuery in package.json and just use a CDN link instead since this is only used for prototyping and examples, not for the library as such.