webpack bundle madness
Splitted webpack.config.js into 4 files:
- webpack.common.js -> using the plugin „webpack-merge“ for shared config files
- webpack.app-dev.js -> for app-development using „webpack-dev-server“
- webpack.app-prod.js -> production of the formform app (app.bundle.js)
- webpack.lib.js -> production of the formform library (formform.dist.js)

The app html and css files now live inside the app/ directory to be more separated from the other stuff. I’m not quite sure yet if this is the optimal organization, since the actual app bundle is missing in app/js/ and only pretends to be there when I’m in development mode (this might be a bit confusing?).

Also, I don’t know if I should separate the app at index.js from all the lib stuff in the build, so that I can just plug in the library and not have everything in one file? I’m not sure how to do this without breaking my modular structure, but I guess it’s not really relevant for now.

Maybe I should also exclude D3 from the app build, since I have done this for the lib, but I guess the app should be more self-contained since it is not something to be distributed like the library. So many questions…