More webpack config, changed export structure
- The lib is now exported to /dist/formform.js, app exports to /app/js/app.bundle.js to keep them separated
- Fixed the defaults problem by adding „libraryExport: 'default‘“ to the lib output so no more formform.defaults now.

For now, I decided to export a single object to the formform namespace with all the modules attached, so there is no privileged module anymore and they can be accessed as „formform.calc“, „formform.form“ (although that looks a bit weird) and „formform.graph“.

This is not ideal, since the library should have a modular import to reduce loading times, but I found that it is the best solution for now, since I may have to restructure the library completely because the whole class inheritence thing seems to be more and more dysfunctional with the further addition of modules.

I’d rather imagine a family or dependency network of modules, where one might import from multiple peer modules that in turn import from other peers. And every module imports everything from a common core that may be fcalc, maybe also some fform. So we have a core api from the common namespace „formform“ that is always present and then users can choose to import other modules, but all their code should still be self-contained, so users shouldn’t have to import other modules for dependencies (I’ll have to find out if import handles redundant code or how to do that). Bundles of all modules as well as the core itself are then tied together in exports.js to be imported as needed. I might just try to do that, let’s hope it works.