Reset to original dna.js from v0.9.2
Previous changes will have to wait for the next gen of formform, where they are more appropriate with what I have in mind for the future of my formula syntax. For now, it’ll remain like this so I can make fixes and integrate feature requests without breaking anything in the meantime.