vmap data structure, refactored save function
Implemented the vmap abstraction/refactorization that I have developed in my earlier (private) Observable notebook. It separates the pure tree structure with position and scaling data from its representation in e.g. SVG format. This enables vmaps to be implemented in different environments or processed as raw data for other purposes.

Each branch of the vmap tree is entirely independent from its parent nodes, harnessing the property of compositionality of vmaps. It also stores additional data that can be re-entered into the vmap function to recreate the exact same (sub-)tree. This will make it easier to analyze and filter vmaps or to create tools for composition and modification.

I have also refactored the save/export function for rendered images so that it can be used from other formform modules (not only graph).

The next thing I’ll do is implement the HTML structure around vmaps (captions, grids) as SVG to be able to export vmaps as images.