pure SVG output for vmap_svg and vmapPerspectives_svg
vmap_svg and vmapPerspectives_svg now produce pure SVG output that can be used outside of HTML for images, PDFs, etc. vmapList_svg will follow. I want to keep the mixed HTML/SVG output available, so this code has to be adapted to the new parameters.

Also integrated my css/styles pattern for dna-svg.js for better separation of style and markup, but I’m not sure how far I should take this here in terms of style classes.

I have also set the output of each svg functions to an object that contains the original vmap tree(s) that has/have been generated and of course the svg element. I don’t know how useful this will be, but I’d like users to have access to the data structure for the purpose of further analysis and processing.

I wonder if I should separate the functions instead though, which would mean more work on behalf of the user, but would also make it clear what output you can expect. Have to rethink this.

I want to get rid of the classes in graph-d3.js and instantiation in fgraph.js in the near future since they don’t seem to be useful after all and we just have to deal with more stateful objects, which is bad. D3 uses the DOM to attach data/state to, so there is really no reason to use OOP here.