Tricorder vmap zoom controls + export png/svg
vmaps can now have padding and background-color for the map itself (vmapPad, vmapC) and for the figure (figPad, figC). This yields more flexibility for the pure SVG output. vmapPerspectives also has individual figPad and figC, but via localOptions can set them for the individual vmaps too.

Still missing is vmapList function in dna-svg.js, it has to be adjusted to the new structure and data input.

Not sure if I should keep the old svg+html („mixed“) output for vmaps, would have to be adjusted but may be not too difficult. Might be helpful to generate vmaps more responsively in notebooks.

FORM tricorder now has zoom controls and functional png/svg export. The blurry output was due to retina pixel scaling, which is especially apparent for very small image sizes. Hence I’ve optimized the vmap output size to compensate. I also improved on the perspectives button and made the background checkbox operate on the svg itself instead of having to re-generate the whole vmap.