New (experimental) formula syntax for re-entries
This new syntax is an attempt of a more iconic, yet clear and easy to write formula notation for self-equivalent re-entry FORMs. The point of re-entry is marked with an “@”, while dots around this symbol indicate the re/pre-entry iterations in order to obtain the self-equivalence. An underscore is used to indicate the first (unmarked) iteration for the open re-entry types.

Definitions:
- `.` := one (p)re-entry step
- `<b>@<a>` := point of re-entry from pre-entered step(s) <a> to re-entered step(s) <b>
- `_` := unmarked pre-entry step (uncounted, always first)

All 6 types of self-equivalent re-entry (as described in uFORMiFORM):
- `@ f` → re-enter `f`
    ⇔ closed FORM with odd/even re-entry number
- `..@ f` → re-enter 2 iterations of `f`
    ⇔ closed FORM with even re-entry number
- `..@. f` → 1 iteration of `f` before re-entering 2 iterations
    ⇔ closed FORM with odd re-entry number
- `@_ f` → 1 unmarked iteration of `f` before re-entry
    ⇔ open FORM with odd/even re-entry number
- `..@_ f` → 1 unmarked iteration of `f` before re-entering 2 iterations
    ⇔ open FORM with even re-entry number
- `..@._ f` → 1 unmarked iteration + 1 iteration of `f` before re-entering 2 iterations
    ⇔ open FORM with odd re-entry number