Separated vmap-svg, fixed inconsistent indentation
The dna-svg.js module is no longer a dependency in formform. It can be installed separately and used in conjunction with formform.dna functions. See new repository at: https://github.com/formsandlines/vmap-svg

Next up is dropping dependency on graph-d3.js, which is a little more involved but this will make the core library much lighter and will also get rid of the D3 dependency.