I successfully splitted formform into 3 repositories by its 3 branches: - `master` is now here: https://github.com/formsandlines/formform-js - `rescript` is now in https://github.com/formsandlines/formform-rescript - `clojure` remains in https://github.com/formsandlines/formform
I actually renamed this repository from `formform` to `formform-js`, so
forks and stars that originally related to this older version still
relate to it and not to one of the rewrites.

For `rescript` and `clojure` I had to cut-off earlier commits that did
not belong to the respective branches by rebasing onto a new orphan
branch. While I kept all commit dates, the SHAs are different now, but
that is okay, since nobody really pulled from those branches anyway.
I should have created those separate repos right from the start to save
me the trouble, since their histories are completely unrelated and they
are actually different projects after all, but I am glad the split went
so smoothly and without any conflicts.

I also updated the README with some infos about the state of this repository.